+++ src/index.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import ReactDOM from "react-dom/client";
 
 import App from "./App.jsx";
@@ -5,3 +6,4 @@ import "./index.css";
 
 const entryPoint = document.getElementById("root");
 ReactDOM.createRoot(entryPoint).render(<App />);
+//ReactDOM.createRoot(entryPoint).render(React.createElement(App))
