+++ vite.config.js
@@ -5,6 +5,6 @@ import react from "@vitejs/plugin-react";
 export default defineConfig({
   plugins: [react()],
   server: {
-    allowedHosts: ["3wfk2w-5174.csb.app", "3wfk2w-5173.csb.app"],
+    allowedHosts: ["3wfk2w-5174.csb.app", "xhrpy5-5173.csb.app"],
   },
 });
